reorganize: collection all invite functionality in one folder
