+++ src/packages/user/user/collection/action/manifests.ts
@@ -1,5 +1,4 @@
 import { UmbCreateUserCollectionAction } from './create-user.collection-action.js';
-import { UmbInviteUserCollectionAction } from './invite-user.collection-action.js';
 import { UMB_COLLECTION_ALIAS_CONDITION } from '@umbraco-cms/backoffice/collection';
 import { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
@@ -21,22 +20,4 @@ export const createManifest: ManifestTypes = {
 	],
 };
 
-export const inviteManifest: ManifestTypes = {
-	type: 'collectionAction',
-	kind: 'button',
-	name: 'Invite User Collection Action',
-	alias: 'Umb.CollectionAction.User.Invite',
-	api: UmbInviteUserCollectionAction,
-	weight: 100,
-	meta: {
-		label: 'Invite',
-	},
-	conditions: [
-		{
-			alias: UMB_COLLECTION_ALIAS_CONDITION,
-			match: 'Umb.Collection.User',
-		},
-	],
-};
-
-export const manifests = [createManifest, inviteManifest];
+export const manifests = [createManifest];
