+++ src/packages/user/user/index.ts
@@ -1,3 +1,4 @@
 export * from './components/index.js';
+export * from './invite/index.js';
 export * from './repository/index.js';
 export * from './types.js';
