+++ src/packages/user/user/invite/manifests.ts
@@ -1,0 +1,5 @@
+import { manifests as collectionActionManifests } from './collection-action/manifests.js';
+import { manifests as modalManifests } from './modal/manifests.js';
+import { manifests as repositoryManifests } from './repository/manifests.js';
+
+export const manifests = [collectionActionManifests, modalManifests, repositoryManifests];
