+++ src/packages/user/user/invite/modal/index.ts
@@ -1,0 +1,2 @@
+export { UMB_INVITE_USER_MODAL } from './invite/index.js';
+export { UMB_RESEND_INVITE_TO_USER_MODAL } from './resend-invite/index.js';
