+++ src/packages/user/user/invite/modal/invite/user-invite-modal.element.ts
@@ -1,5 +1,5 @@
-import { UmbUserGroupInputElement } from '../../../user-group/components/input-user-group/user-group-input.element.js';
-import { UmbInviteUserRepository } from '../../repository/invite/invite-user.repository.js';
+import { UmbUserGroupInputElement } from '../../../../user-group/components/input-user-group/user-group-input.element.js';
+import { UmbInviteUserRepository } from '../../repository/invite-user.repository.js';
 import { css, html, customElement, query } from '@umbraco-cms/backoffice/external/lit';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { UmbModalBaseElement } from '@umbraco-cms/backoffice/modal';
