+++ src/packages/user/user/invite/modal/resend-invite/resend-invite-to-user-modal.element.ts
@@ -1,4 +1,4 @@
-import { UmbInviteUserRepository } from '../../repository/invite/invite-user.repository.js';
+import { UmbInviteUserRepository } from '../../repository/invite-user.repository.js';
 import { css, html, customElement } from '@umbraco-cms/backoffice/external/lit';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { UmbModalBaseElement } from '@umbraco-cms/backoffice/modal';
