+++ src/packages/user/user/invite/repository/index.ts
@@ -1,0 +1,2 @@
+export { UmbInviteUserRepository } from './invite-user.repository.js';
+export { UMB_INVITE_USER_REPOSITORY_ALIAS } from './manifests.js';
