+++ src/packages/user/user/invite/repository/invite-user.repository.ts
@@ -1,4 +1,4 @@
-import { UmbUserRepositoryBase } from '../user-repository-base.js';
+import { UmbUserRepositoryBase } from '../../repository/user-repository-base.js';
 import { type UmbInviteUserDataSource } from './types.js';
 import { UmbInviteUserServerDataSource } from './invite-user.server.data-source.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
