+++ src/packages/user/user/manifests.ts
@@ -1,4 +1,5 @@
 import { manifests as collectionManifests } from './collection/manifests.js';
+import { manifests as inviteManifests } from './invite/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as workspaceManifests } from './workspace/manifests.js';
 import { manifests as modalManifests } from './modals/manifests.js';
@@ -9,6 +10,7 @@ import { manifests as conditionsManifests } from './conditions/manifests.js';
 
 export const manifests = [
 	...collectionManifests,
+	...inviteManifests,
 	...repositoryManifests,
 	...workspaceManifests,
 	...modalManifests,
