--- src/packages/user/user/modals/index.ts
@@ -1,2 +1,0 @@
-export { UMB_RESEND_INVITE_TO_USER_MODAL } from './resend-invite/index.js';
-export { UMB_INVITE_USER_MODAL } from './invite/index.js';
