+++ src/packages/user/user/repository/index.ts
@@ -2,7 +2,6 @@ export * from './change-password/change-user-password.repository.js';
 export * from './disable/disable-user.repository.js';
 export * from './enable/enable-user.repository.js';
 export * from './unlock/unlock-user.repository.js';
-export * from './invite/invite-user.repository.js';
 export * from './user.repository.js';
 export * from './item/index.js';
 export * from './detail/index.js';
