+++ src/packages/user/user/repository/manifests.ts
@@ -4,7 +4,6 @@ import { UmbDisableUserRepository } from './disable/disable-user.repository.js';
 import { UmbEnableUserRepository } from './enable/enable-user.repository.js';
 import { UmbChangeUserPasswordRepository } from './change-password/change-user-password.repository.js';
 import { UmbUnlockUserRepository } from './unlock/unlock-user.repository.js';
-import { UmbInviteUserRepository } from './invite/invite-user.repository.js';
 import { manifests as itemManifests } from './item/manifests.js';
 import type { ManifestStore, ManifestRepository } from '@umbraco-cms/backoffice/extension-registry';
 
@@ -48,14 +47,6 @@ const unlockRepository: ManifestRepository = {
 	api: UmbUnlockUserRepository,
 };
 
-export const UMB_INVITE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Invite';
-const inviteRepository: ManifestRepository = {
-	type: 'repository',
-	alias: UMB_INVITE_USER_REPOSITORY_ALIAS,
-	name: 'Invite User Repository',
-	api: UmbInviteUserRepository,
-};
-
 const store: ManifestStore = {
 	type: 'store',
 	alias: 'Umb.Store.User',
@@ -69,7 +60,6 @@ export const manifests = [
 	enableRepository,
 	changePasswordRepository,
 	unlockRepository,
-	inviteRepository,
 	store,
 	...itemManifests,
 ];
